moved the upCase/downCase helpers into a separate file
