--- memef/memef.c
@@ -1,13 +1,0 @@
-#include <stdmeme.h>
-#include <stdarg.h>
-
-
-void memef(char *pat, ...)
-{
-
-
-
-
-
-
-}
