--- memef/stdmeme.h
@@ -1,11 +1,0 @@
-#include <string.h>
-#include <stdlib.h>
-#include <stdio.h>
-
-void memef(char *pat, ...);
-
-char *spongeCase(char *str); //SpOnGeCaSe
-char *owOofCase(char *str);  //_o_*w*_O_*o*_f_*C*_a_*s*_e_
-char *aestheticCase(char *str);  // a e s t h e t i c C a s e
-
-
