--- memef/test.c
@@ -1,6 +1,0 @@
-#include "stdmeme.h"
-
-int main(){
-    printf("%s\n", spongeCase("This is a test"));
-    return 0;
-}
