+++ stdmeme/stdmeme.h
@@ -2,10 +2,18 @@
 #include <stdlib.h>
 #include <stdio.h>
 
+
+/*
+ * The Good Stuff A++
+ */
 void memef(char *pat, ...);
 
 char *spongeCase(char *str); //SpOnGeCaSe
 char *owOofCase(char *str);  //_o_*w*_O_*o*_f_*C*_a_*s*_e_
 char *aestheticCase(char *str);  // a e s t h e t i c C a s e
 
-
+/*
+ * Some helpers for the good stuff
+ */
+char upCase(char c);
+char downCase(char c);
